Moved Modals to Templates/Modals
Template partials should not be in apps include folder
