+++ templates/modals/delete_prompt.php
