+++ templates/modals/switch_pool.php
